Move JdbcUtils and RandomDataTestUtils back to main, so they can be shared among modules; put testing helper classes in a separate package
