+++ base/src/main/java/uk/ac/ebi/atlas/testutils/RandomDataTestUtils.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas;
+package uk.ac.ebi.atlas.testutils;
 
 import com.google.common.collect.ImmutableList;
