+++ base/src/test/java/uk/ac/ebi/atlas/experimentimport/analytics/singlecell/clusters/ClustersDaoIT.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
-import uk.ac.ebi.atlas.RandomDataTestUtils;
+import uk.ac.ebi.atlas.testutils.RandomDataTestUtils;
 
 import javax.inject.Inject;
 import java.util.List;
