+++ base/src/test/java/uk/ac/ebi/atlas/experimentimport/analytics/singlecell/tsne/TSnePlotStreamerIT.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.ac.ebi.atlas.JdbcUtils;
+import uk.ac.ebi.atlas.testutils.JdbcUtils;
 import uk.ac.ebi.atlas.resource.DataFileHub;
 import uk.ac.ebi.atlas.resource.DataFileHubFactory;
