+++ base/src/test/java/uk/ac/ebi/atlas/profiles/stream/RnaSeqProfileStreamFactoryTest.java
@@ -8,7 +8,7 @@ import uk.ac.ebi.atlas.model.AssayGroup;
 import uk.ac.ebi.atlas.model.experiment.differential.*;
 import uk.ac.ebi.atlas.model.experiment.differential.rnaseq.RnaSeqProfile;
 import uk.ac.ebi.atlas.profiles.IterableObjectInputStream;
-import uk.ac.ebi.atlas.resource.MockDataFileHub;
+import uk.ac.ebi.atlas.testutils.MockDataFileHub;
 import uk.ac.ebi.atlas.web.DifferentialRequestPreferences;
 
 import java.util.Collection;
