+++ base/src/test/java/uk/ac/ebi/atlas/trader/ConfigurationTraderTest.java
@@ -8,10 +8,9 @@ import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.ac.ebi.atlas.model.experiment.baseline.BaselineExperimentConfiguration;
 import uk.ac.ebi.atlas.model.experiment.baseline.Factor;
-import uk.ac.ebi.atlas.resource.MockDataFileHub;
+import uk.ac.ebi.atlas.testutils.MockDataFileHub;
 
 import java.text.MessageFormat;
-import java.util.List;
 import java.util.Map;
 
 import static org.hamcrest.Matchers.is;
